feat(esp8266): support 64bit sleep time for esp_deep_sleep

See merge request sdk/ESP8266_RTOS_SDK!1516

+++ components/esp8266/include/esp_sleep.h
@@ -65,7 +65,7 @@ typedef enum {
   *
   * @return    null
   */
-void esp_deep_sleep(uint32_t time_in_us);
+void esp_deep_sleep(uint64_t time_in_us);
 
 /**
  * @brief Set implementation-specific power management configuration
